Renamed rz_flirt to rz_sign & moved sign.c in it
