+++ librz/analysis/analysis.c
@@ -41,28 +41,6 @@ static void meta_count_for(RzEvent *ev, int type, void *user, void *data) {
 	se->res = rz_meta_space_count_for(analysis, se->data.count.space);
 }
 
-static void zign_unset_for(RzEvent *ev, int type, void *user, void *data) {
-	RzSpaces *s = (RzSpaces *)ev->user;
-	RzAnalysis *analysis = container_of(s, RzAnalysis, zign_spaces);
-	RzSpaceEvent *se = (RzSpaceEvent *)data;
-	rz_sign_space_unset_for(analysis, se->data.unset.space);
-}
-
-static void zign_count_for(RzEvent *ev, int type, void *user, void *data) {
-	RzSpaces *s = (RzSpaces *)ev->user;
-	RzAnalysis *analysis = container_of(s, RzAnalysis, zign_spaces);
-	RzSpaceEvent *se = (RzSpaceEvent *)data;
-	se->res = rz_sign_space_count_for(analysis, se->data.count.space);
-}
-
-static void zign_rename_for(RzEvent *ev, int type, void *user, void *data) {
-	RzSpaces *s = (RzSpaces *)ev->user;
-	RzAnalysis *analysis = container_of(s, RzAnalysis, zign_spaces);
-	RzSpaceEvent *se = (RzSpaceEvent *)data;
-	rz_sign_space_rename_for(analysis, se->data.rename.space,
-		se->data.rename.oldname, se->data.rename.newname);
-}
-
 void rz_analysis_hint_storage_init(RzAnalysis *a);
 
 void rz_analysis_hint_storage_fini(RzAnalysis *a);
@@ -110,10 +88,6 @@ RZ_API RzAnalysis *rz_analysis_new(void) {
 	rz_event_hook(analysis->meta_spaces.event, RZ_SPACE_EVENT_UNSET, meta_unset_for, NULL);
 	rz_event_hook(analysis->meta_spaces.event, RZ_SPACE_EVENT_COUNT, meta_count_for, NULL);
 
-	rz_spaces_init(&analysis->zign_spaces, "zs");
-	rz_event_hook(analysis->zign_spaces.event, RZ_SPACE_EVENT_UNSET, zign_unset_for, NULL);
-	rz_event_hook(analysis->zign_spaces.event, RZ_SPACE_EVENT_COUNT, zign_count_for, NULL);
-	rz_event_hook(analysis->zign_spaces.event, RZ_SPACE_EVENT_RENAME, zign_rename_for, NULL);
 	rz_analysis_hint_storage_init(analysis);
 	rz_interval_tree_init(&analysis->meta, rz_meta_item_free);
 	analysis->typedb = rz_type_db_new();
