+++ librz/include/rz_flirt.h
@@ -13,8 +13,6 @@
 extern "C" {
 #endif
 
-RZ_LIB_VERSION_HEADER(rz_flirt);
-
 #define RZ_FLIRT_NAME_MAX 1024
 
 typedef struct RzFlirtTailByte {
