+++ librz/main/rizin.c
@@ -60,7 +60,6 @@ static int rz_main_version_verify(int show) {
 		{ "rz_syscall", rz_syscall_version },
 		{ "rz_util", rz_util_version },
 		{ "rz_diff", rz_diff_version },
-		{ "rz_flirt", rz_flirt_version },
 		{ "rz_demangler", rz_demangler_version },
 		/* ... */
 		{ NULL, NULL }
