+++ librz/signature/sign.c
