+++ test/unit/test_sign.c
@@ -9,6 +9,7 @@
 
 static bool test_analysis_sign_get_set(void) {
 	RzAnalysis *analysis = rz_analysis_new();
+	rz_sign_analysis_set_hooks(analysis);
 
 	RzSignItem *item = rz_sign_item_new();
 	item->name = strdup("sym.mahboi");
